requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0,<2.0.0
faster-whisper>=1.0.0
pyobjc-core>=10.0
pyobjc-framework-Quartz>=10.0
pyobjc-framework-Cocoa>=10.0
